fixed rounding error for determinants

+++ include/determinant.h
@@ -1,2 +1,2 @@
-int determinant_recur(Matrix *A);
-int determinant_lu(Matrix *A);
+double determinant_recur(Matrix *A);
+double determinant_lu(Matrix *A);
